apps/api/requirements.txt
azure-functions==1.24.0
MarkupSafe==3.0.3
pyodbc==5.2.0
Werkzeug==3.1.3
